pyproject.toml
[project]
name = "rag-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
